+++ cli/sous_query_gdm.go
@@ -8,6 +8,7 @@ import (
 	"github.com/opentable/sous/lib"
 	"github.com/opentable/sous/util/cmdr"
 	"github.com/opentable/sous/util/logging"
+	"github.com/opentable/sous/util/logging/messages"
 )
 
 // SousQueryGDM is the description of the `sous query gdm` command
@@ -37,7 +38,7 @@ func (*SousQueryGDM) RegisterOn(psy Addable) {
 
 // Execute defines the behavior of `sous query gdm`
 func (sb *SousQueryGDM) Execute(args []string) cmdr.Result {
-	logging.Log.Vomit.Printf("%v", sb.GDM.Snapshot())
+	messages.ReportLogFieldsMessage("snapshot", logging.ExtraDebug1Level, logging.Log, sb.GDM.Snapshot())
 	sous.DumpDeployments(os.Stdout, sb.GDM.Deployments)
 	return cmdr.Success()
 }